poker-agent-arena/backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-socketio==5.11.0

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# Solana
solana==0.32.0
solders==0.20.0
base58==2.1.1

# Anthropic AI
anthropic==0.18.1

# Cryptography
cryptography==42.0.2
pynacl==1.5.0

# Settings & Validation
pydantic==2.6.1
pydantic-settings==2.2.1

# Utilities
python-dotenv==1.0.1
httpx==0.26.0

# Development
ruff==0.2.2
pytest==8.0.1
pytest-asyncio==0.23.4
pytest-cov==4.1.0
